Move the dashboard's load layer into a hook of its own
`ProjectList.tsx` went back over the 900-LOC ceiling adding the folder
timeouts. Rather than raise the limit, the loading concern comes out whole:
state, the sequence token, both fetches, `loadAll`, and the mount/workspace
effect. 917 -> 777.

It is the right seam rather than a convenient one. This is the part that keeps
being wrong — it has produced an unbounded spinner twice, the second time
because only one of the two calls `loadAll` awaits had a timeout. In a file of
its own both bounds sit next to each other, where it is visible that they have
to match.

The setters come back out of the hook on purpose: renames and deletions edit
the list in place rather than re-scanning the whole projects folder.

No behaviour change. The four ProjectList tests, which drive this whole path,
pass unchanged.
